+++ clients/admin-ui/cypress/support/stubs.ts
@@ -566,6 +566,15 @@ export const stubSystemAssets = () => {
   cy.intercept("GET", "/api/v1/plus/system-assets/*", {
     fixture: "systems/system_assets",
   }).as("getSystemAssets");
+  cy.intercept("POST", "/api/v1/plus/system-assets/*/assets", {
+    response: 200,
+  }).as("addSystemAsset");
+  cy.intercept("PUT", "/api/v1/plus/system-assets/*/assets*", {
+    response: 200,
+  }).as("updateSystemAssets");
+  cy.intercept("DELETE", "/api/v1/plus/system-assets/*/assets*", {
+    response: 200,
+  }).as("deleteSystemAssets");
 };
 
 export const stubDataCatalog = () => {